+++ tests/testthat/test-translate.r
@@ -75,8 +75,8 @@ test_that("n_distinct can take multiple values", {
   )
 })
 
-test_that("na_if is translated to NULL_IF", {
-  expect_equal(translate_sql(na_if(x, 0L)), sql('NULL_IF("x", 0)'))
+test_that("na_if is translated to NULLIF (#211)", {
+  expect_equal(translate_sql(na_if(x, 0L)), sql('NULLIF("x", 0)'))
 })
 
 test_that("connection affects quoting character", {